Added rotation and initial condition
I added the initial condition of u (body frame x velocity) equal to the
airspeed (5). I also added a rotation of the wind speed vectors into/out
of the NED and Body Frames. Not sure if it is fully working though. The
lack of deltas showing up on the graphs is normal at this stage because
the values are not yet ported through to the display.